requirements.txt
torch>=2.0.0
torchvision>=0.15.0
numpy>=1.21.0
pillow>=8.0.0
wandb>=0.15.0
tqdm>=4.62.0
timm>=0.9.0
pyyaml>=6.0
matplotlib>=3.5.0
seaborn>=0.11.0
tensorboard>=2.10.0
scipy>=1.9.0